Clean up mac warnings

+++ test/TestRle.cc
@@ -119,7 +119,8 @@ TEST(RLEv1, testSigned) {
   }
   rle->next(data.data(), 30, nullptr);
   for(size_t i = 0; i < 30; ++i) {
-    EXPECT_EQ(16 - 100 - i, data[i]) << "Wrong output at " << (i + 100);
+    EXPECT_EQ(16 - 100 - static_cast<long>(i), data[i]) 
+      << "Wrong output at " << (i + 100);
   }
 }
